pyproject.toml
[project]
name = "niceapimanager"
version = "0.1.0"
description = "NiceApiManager integrated backend and React admin console."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.15.2",
    "apscheduler==3.10.4",
    "fastapi==0.115.12",
    "httpx==0.28.1",
    "pydantic-settings==2.8.1",
    "sqlalchemy==2.0.39",
    "uvicorn[standard]==0.34.0",
]

[tool.uv]
package = false

